better looking display of frame gallery
