+++ src/javascript/app/components/Frame/index.jsx
@@ -1,44 +1,40 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import GameBoyImage from '../GameBoyImage/component';
-import applyFrame from '../../../tools/applyFrame';
-import textToTiles from '../../../tools/textToTiles';
+import './index.scss';
+import useFrame from './useFrame';
 
-const getTiles = ({ id, name }) => {
-  const text = `\n  ID: ${id}\n\n  ${name}`;
-  return applyFrame(textToTiles(text), id);
-};
 
 const Frame = ({ id, name, palette }) => {
-  const [tiles, setTiles] = useState(null);
-
-  useEffect(() => {
-    let mounted = true;
-    getTiles({
-      id,
-      name,
-    }).then((t) => {
-      if (mounted) {
-        setTiles(t);
-      }
-    });
-
-    return () => {
-      mounted = false;
-    };
-  }, [id, name]);
+  const { deleteFrame, tiles } = useFrame({ id, name });
 
   if (!tiles) {
     return null;
   }
 
   return (
-    <GameBoyImage
-      lockFrame={false}
-      invertPalette={false}
-      palette={palette}
-      tiles={tiles}
-    />
+    <li className="frame">
+      <div className="frame__image">
+        <GameBoyImage
+          lockFrame={false}
+          invertPalette={false}
+          palette={palette}
+          tiles={tiles}
+        />
+      </div>
+
+      <button
+        type="button"
+        onClick={() => {
+          // eslint-disable-next-line no-alert
+          if (window.confirm('really delete?')) {
+            deleteFrame(id);
+          }
+        }}
+      >
+        DELETE!!
+      </button>
+    </li>
   );
 };
