+++ src/javascript/app/components/Frames/index.scss
@@ -3,12 +3,7 @@
   &__list {
     display: flex;
     flex-wrap: wrap;
-    margin: 0 -8px;
-    justify-content: space-between;
+    justify-content: center;
+    margin: 0 -5px 20px;
   }
-}
-
-.frame {
-  width: 160px;
-  margin: 0 8px 16px 8px;
 }
